create authority pipeline

+++ wikia_dstk/pipeline/data_extraction/run.py
@@ -5,38 +5,51 @@ listed in order to warm the cache.
 """
 
 import re
-import sys
 from boto import connect_s3
 from boto.ec2 import connect_to_region
 from boto.utils import get_instance_metadata
 from subprocess import Popen
 from time import sleep
+from config import default_config
+from ... import get_argparser_from_config, argstring_from_namespace
 
-from config import config
 
-workers = int(sys.argv[1])
-BUCKET = connect_s3().get_bucket('nlp-data')
+def get_args():
+    ap = get_argparser_from_config(default_config)
+    ap.add_argument('--workers', dest='workers', type=int, default=8)
+    ap.add_argument('--no-shutdown', dest='do_shutdown', type='store_false', default=True)
+    return ap.parse_known_args()
 
-counter = 0
-while True:
-    processes = []
-    keys = [key.name for key in BUCKET.list(prefix='data_events/') if
-            re.sub(r'/?data_events/?', '', key.name) is not '']
-    while len(keys) > 0:
-        counter = 0
-        while len(processes) < workers:
-            processes.append(
-                Popen(
-                    '/home/ubuntu/venv/bin/python -m ' +
-                    'wikia_dstk.pipeline.data_extraction.child %s' % keys.pop(),
-                    shell=True))
-        processes = filter(lambda x: x.poll() is None, processes)
-        sleep(0.25)
-    counter += 1
-    print 'No more keys, waiting 15 seconds. Counter: %d/20' % counter
-    if counter >= 20:
-        print 'Scaling down, shutting down.'
-        current_id = get_instance_metadata()['instance-id']
-        ec2_conn = connect_to_region(config['region'])
-        ec2_conn.terminate_instances([current_id])
-    sleep(15)
+
+def main():
+    args, extras = get_args()
+    bucket = connect_s3().get_bucket('nlp-data')
+
+    counter = 0
+    while True:
+        processes = []
+        keys = [key.name for key in bucket.list(prefix='%s/' % args.queue) if
+                re.sub(r'/?%s/?' % args.queue, '', key.name) is not '']
+        while len(keys) > 0:
+            counter = 0
+            while len(processes) < args.workers:
+                processes.append(
+                    Popen(
+                        '/home/ubuntu/venv/bin/python -m ' +
+                        'wikia_dstk.pipeline.data_extraction.child ' +
+                        '%s --key=%s' % (argstring_from_namespace(args, extras), keys.pop()),
+                        shell=True))
+            processes = filter(lambda x: x.poll() is None, processes)
+            sleep(0.25)
+        counter += 1
+        print 'No more keys, waiting 15 seconds. Counter: %d/20' % counter
+        if args.do_shutdown and counter >= 20:
+            print 'Scaling down, shutting down.'
+            current_id = get_instance_metadata()['instance-id']
+            ec2_conn = connect_to_region(args.region)
+            ec2_conn.terminate_instances([current_id])
+        sleep(15)
+
+
+if __name__ == '__main__':
+    main()
